pyproject.toml
[project]
name = "admet-dkl"
version = "0.1.0"
description = "Benchmarking SV-DKL vs MLP+MC-Dropout on TDC ADMET with MoLFormer embeddings"
requires-python = "==3.11.*"
dependencies = [
    "torch>=2.4.0",
    "gpytorch>=1.13",
    "linear-operator>=0.5",
    "transformers>=4.40",
    "PyTDC>=0.4.1",
    "hydra-core>=1.3",
    "optuna>3.0.0",
    "pandas>=2.2",
    "scikit-learn>=1.5",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "tqdm>=4.66",
    "scipy>=1.13",
    "tensorboard>=2.20.0",
    "hydra-optuna-sweeper<1.2.0",
    "xgboost>=2.0",
    "sqlalchemy<2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-xdist>=3.5",
]

[tool.uv]
# tiledbsoma (pulled in by pytdc) fails to build on Windows; it is only used
# by TDC's CellXGene benchmark — not needed for ADMET. Exclude via impossible marker.
override-dependencies = [
    "tiledbsoma; python_version < '0'",
]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
